+++ spec/factories/paper_tracker_query_factory.rb
@@ -1,0 +1,11 @@
+FactoryGirl.define do
+  factory :paper_tracker_query do
+    sequence :title do |n|
+      "Tracker Query ##{n}"
+    end
+
+    sequence :query do |n|
+      "TASK authors COMPLETED > #{n} DAYS"
+    end
+  end
+end