+++ CHANGELOG.md
@@ -4,7 +4,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 
 ## [Unreleased][unreleased]
 ### Added
--
+- Paper tracker: shared saved searches
 
 ### Changed
 -